feat: add share utility and provider
