+++ app/src/main/res/xml/file_paths.xml
@@ -1,0 +1,3 @@
+<paths xmlns:android="http://schemas.android.com/apk/res/android">
+    <cache-path name="composereports" path="composereports/" />
+</paths>
